+++ core/model/application.go
@@ -86,6 +86,16 @@ type AppOrgRole struct {
 	DateUpdated *time.Time
 }
 
+//GetPermissionNamed returns the permission for a name if the role has it
+func (c AppOrgRole) GetPermissionNamed(name string) *Permission {
+	for _, permission := range c.Permissions {
+		if permission.Name == name {
+			return &permission
+		}
+	}
+	return nil
+}
+
 //CheckAssigners checks if the passed permissions satisfy the needed assigners for all role permissions
 func (c AppOrgRole) CheckAssigners(assignerPermissions []string) error {
 	if len(c.Permissions) == 0 {
@@ -120,6 +130,30 @@ type AppOrgGroup struct {
 
 	DateCreated time.Time
 	DateUpdated *time.Time
+}
+
+//CheckAssigners checks if the passed permissions satisfy the needed assigners for the group
+func (cg AppOrgGroup) CheckAssigners(assignerPermissions []string) error {
+	//check permission
+	if len(cg.Permissions) > 0 {
+		for _, permission := range cg.Permissions {
+			err := permission.CheckAssigners(assignerPermissions)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	//check roles
+	if len(cg.Roles) > 0 {
+		for _, role := range cg.Roles {
+			err := role.CheckAssigners(assignerPermissions)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	//all assigners are satisfied
+	return nil
 }
 
 func (cg AppOrgGroup) String() string {
@@ -230,7 +264,8 @@ func (ao ApplicationOrganization) IsAuthTypeSupported(appType ApplicationType, a
 type IdentityProviderSetting struct {
 	IdentityProviderID string `bson:"identity_provider_id"`
 
-	UserIdentifierField string `bson:"user_identifier_field"`
+	UserIdentifierField string            `bson:"user_identifier_field"`
+	ExternalIDFields    map[string]string `bson:"external_id_fields"`
 
 	FirstNameField  string `bson:"first_name_field"`
 	MiddleNameField string `bson:"middle_name_field"`