+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Phone password auth type [#434](https://github.com/rokwire/core-building-block/issues/434)
 
 
+## [1.22.0] - 2022-04-02
+### Added
+- Expose revoke roles from account Admin API [#412](https://github.com/rokwire/core-building-block/issues/412)
+- Expose revoke permissions from account Admin API [#411](https://github.com/rokwire/core-building-block/issues/411)
+- Expose grant permissions to role Admin API [#415](https://github.com/rokwire/core-building-block/issues/415)
+- Expose remove accounts from a group Admin API [#413](https://github.com/rokwire/core-building-block/issues/413)
+- Expose add accounts to a group Admin API [#384](https://github.com/rokwire/core-building-block/issues/384)
+- Handle external ID management [#364](https://github.com/rokwire/core-building-block/issues/364)
+
 ## [1.21.1] - 2022-03-17
 ### Fixed
 - Fix verify credential HTML template loading issues [#451](https://github.com/rokwire/core-building-block/issues/451)
@@ -258,8 +267,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.21.1...HEAD
-[1.21.0]: https://github.com/rokwire/core-building-block/compare/v1.21.0...v1.21.1
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.22.0...HEAD
+[1.22.0]: https://github.com/rokwire/core-building-block/compare/v1.21.1...v1.22.0
+[1.21.1]: https://github.com/rokwire/core-building-block/compare/v1.21.0...v1.21.1
 [1.21.0]: https://github.com/rokwire/core-building-block/compare/v1.20.1...v1.21.0
 [1.20.1]: https://github.com/rokwire/core-building-block/compare/v1.20.0...v1.20.1
 [1.20.0]: https://github.com/rokwire/core-building-block/compare/v1.19.0...v1.20.0